Backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.0
dj-database-url>=2.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pytest>=8.3.0
pytest-django>=4.9.0
pytest-cov>=5.0.0
